requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
statsmodels
imbalanced-learn
xgboost 
lightgbm
boto3 
python-dotenv
mlflow